backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2023.11.16
whisper-openai==1.1.10
python-multipart==0.0.6
torch==2.1.0
torchaudio==2.1.0
langchain==0.0.349
faiss-cpu==1.7.4
requests==2.31.0
openai==1.3.0  # Needed for DeepSeek API compatibility